load data in ArrayDequeImpl
