+++ jraft-core/src/main/java/com/alipay/sofa/jraft/storage/log/SkipListLogStorage.java
@@ -1,0 +1,13 @@
+package com.alipay.sofa.jraft.storage.log;
+
+import com.alipay.sofa.jraft.storage.LogStorage;
+import com.alipay.sofa.jraft.util.Describer;
+
+/**
+ * @author: caobiao
+ * @date: 2020/07/31
+ * @Description:
+ */
+//public class SkipListLogStorage implements LogStorage, Describer {
+//
+//}
